created common reducer
